Refactor command arguments in cloud.py
The update restructures the command-line arguments in the 'cloud.py' script for better clarity and consistency. The option 's3_dir' is relabeled as 'target_dir' for better context and similarly, the option flags for 'num-instances' and 'ec2-type' are simplified for ease of use. Additionally, unnecessary passing of 'obj' in 'delete_endpoint' function has been removed.